backend/requirements.txt
autopep8==1.4.4
Django==2.2.9
django-allauth==0.41.0
django-cors-headers==3.2.1
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
django-rest-auth==0.9.5
Pillow==7.0.0
pylint==2.4.4
pytz==2019.3
